+++ src/main/java/core/mvc/param/extractor/simple/IntParser.java
@@ -1,4 +1,4 @@
-package core.mvc.param.parser.simple;
+package core.mvc.param.extractor.simple;
 
 public class IntParser implements Parser<Integer> {
