+++ src/main/java/core/mvc/param/extractor/simple/LongParser.java
@@ -1,4 +1,4 @@
-package core.mvc.param.parser.simple;
+package core.mvc.param.extractor.simple;
 
 public class LongParser implements Parser<Long> {
